requirements.txt
# requirements.txt
ultralytics
opencv-python
pillow
tkinter
numpy
